+++ apps/web/app/layout.tsx
@@ -28,6 +28,7 @@ import "./v125.css";
 import "./v130.css";
 import "./v131.css";
 import "./v134.css";
+import "./v137.css";
 
 export const metadata: Metadata = {
   title: {